Update joint.m
fix empty window bug

+++ +gfx2d/joint.m
@@ -160,7 +160,9 @@ classdef joint < gfx2d.RigidBody
         
         function delete(obj)
             delete(obj.pl);
-            obj.window.deleteObject(obj.id);
+            if ~isempty(obj.window)
+                obj.window.deleteObject(obj.id);
+            end
         end
         
         function set.color(obj,newcolor)
